more type changes

+++ spuce/filters/create_remez_lpfir.cpp
@@ -13,7 +13,7 @@ void create_remez_lpfir(fir_coeff<float_type>& remezfir, float_type pass_edge, f
   std::vector<float_type> e1(4);
   std::vector<float_type> f1(4);
   std::vector<float_type> w1(4);
-  long nfilt = remezfir.number_of_taps();
+  size_t nfilt = remezfir.number_of_taps();
   remez_fir Remz;
   w1[0] = 1.0;
   w1[1] = stop_weight;
@@ -26,9 +26,9 @@ void create_remez_lpfir(fir_coeff<float_type>& remezfir, float_type pass_edge, f
   std::vector<float_type> fir_coef(nfilt);
   ok = Remz.remez(fir_coef, nfilt, 2, e1, f1, w1, remez_type::BANDPASS);
   if (ok) {
-    for (int i = 0; i < nfilt; i++) remezfir.settap(i, fir_coef[i]);
+    for (size_t i = 0; i < nfilt; i++) remezfir.settap(i, fir_coef[i]);
   } else {
-    for (int i = 0; i < nfilt; i++) remezfir.settap(i, 0);
+    for (size_t i = 0; i < nfilt; i++) remezfir.settap(i, 0);
     remezfir.settap(0, 1);
   }
 }
